create report package
